Add and refactor unit tests for connection handling
Refactored `svc_test.go` by extracting and moving connection handling tests to a new `conn_test.go` file for better modularity and clarity. Introduced new unit tests for `HandleHTTPConnection` behavior and updated `svc_test.go` to focus on service-specific functionality like endpoint generation. This improves test structure and coverage.
